Add node-style build config and examples. Rename abi to json.
