+++ .gitignore
@@ -1,4 +1,5 @@
 dist/
+dist-web/
 node_modules
 package-lock.json
 .DS_Store
