+++ package.json
@@ -2,10 +2,11 @@
   "name": "eosjs2",
   "version": "0.0.0",
   "description": "Talk to eos API",
-  "main": "src/index.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack"
+    "build": "tsc -p ./tsconfig.node.json",
+    "build-web": "webpack"
   },
   "author": "block.one",
   "license": "MIT",
